Added Versioning (#3)
* Added version logic
* Version and translation logic moved to its own extension
